test(e2e): move environment filter pod test to its own spec file
Move the 'Verification of pod filtering by environment' suite out of
pod-smoke.spec.ts into a dedicated pod-environment-filter.spec.ts.
The new file carries its own beforeAll/afterAll lifecycle hooks and
the locator handler that auto-dismisses the async Podman default-machine
dialog. pod-smoke.spec.ts is cleaned of the now-unused imports and
constants.
