+++ tests/playwright/src/specs/pod-smoke.spec.ts
@@ -20,13 +20,9 @@ import * as os from 'node:os';
 
 import { ContainerState, PodState } from '/@/model/core/states';
 import type { ContainerInteractiveParams } from '/@/model/core/types';
-import { CreateMachinePage } from '/@/model/pages/create-machine-page';
 import { PodsPage } from '/@/model/pages/pods-page';
-import { ResourcesPage } from '/@/model/pages/resources-page';
 import { expect as playExpect, test } from '/@/utility/fixtures';
-import { deleteContainer, deleteImage, deletePod, deletePodmanMachine } from '/@/utility/operations';
-import { isLinux } from '/@/utility/platform';
-import { getVirtualizationProvider } from '/@/utility/provider';
+import { deleteContainer, deleteImage, deletePod } from '/@/utility/operations';
 import { waitForPodmanMachineStartup, waitUntil, waitWhile } from '/@/utility/wait';
 
 let backendPort: string;
@@ -43,13 +39,6 @@ const containerNames = ['container1', 'container2', 'container3'];
 const podNames = ['pod1', 'pod2', 'pod3'];
 const containerStartParams: ContainerInteractiveParams = { attachTerminal: false };
 let resetTestData = true;
-
-const secondMachineVisibleName = 'podman-machine-second';
-const secondMachineDisplayName = 'Podman Machine second';
-const defaultMachineDisplayName = 'Podman Machine';
-const environmentFilterImage = 'ghcr.io/linuxcontainers/alpine';
-const environmentFilterContainer = 'environment-filter-container';
-const environmentFilterPod = 'environment-filter-pod';
 
 test.beforeAll(async ({ runner, welcomePage, page, navigationBar }) => {
   runner.setVideoAndTraceName('pods-e2e');
@@ -402,99 +391,5 @@ test.describe
           await playExpect.poll(async () => await podsPage.podExists(pod), { timeout: 15_000 }).toBeFalsy();
         }
       });
-    });
-  });
-
-test.describe
-  .serial('Verification of pod filtering by environment', { tag: '@pdmachine' }, () => {
-    test.skip(
-      isLinux || process.env.TEST_PODMAN_MACHINE !== 'true',
-      'Test suite requires a second Podman machine and should only run when TEST_PODMAN_MACHINE is true',
-    );
-
-    test.beforeAll(async ({ page }) => {
-      // The "set as default?" Podman dialog fires asynchronously when the
-      // second machine finishes starting — potentially in the middle of a
-      // later test. Register a locator handler so Playwright auto-dismisses
-      // it whenever it appears, before it can block any UI interaction.
-      const podmanDialog = page.getByRole('dialog', { name: 'Podman', exact: true });
-      await page.addLocatorHandler(podmanDialog, async () => {
-        await podmanDialog.getByRole('button', { name: 'Ignore' }).click();
-      });
-    });
-
-    test.afterAll(async ({ page }) => {
-      test.setTimeout(120_000);
-
-      try {
-        await deletePod(page, environmentFilterPod);
-        await deleteContainer(page, environmentFilterContainer);
-        await deleteImage(page, environmentFilterImage);
-      } finally {
-        await deletePodmanMachine(page, secondMachineVisibleName);
-      }
-    });
-
-    test('Creating a second Podman machine', async ({ navigationBar, page }) => {
-      test.setTimeout(200_000);
-
-      const settingsBar = await navigationBar.openSettings();
-      await settingsBar.resourcesTab.click();
-      const resourcesPage = new ResourcesPage(page);
-      await playExpect(resourcesPage.heading).toBeVisible();
-      await playExpect.poll(async () => await resourcesPage.resourceCardIsVisible('podman')).toBeTruthy();
-      await resourcesPage.goToCreateNewResourcePage('podman');
-
-      const createMachinePage = new CreateMachinePage(page);
-      await createMachinePage.createMachine(secondMachineVisibleName, {
-        isRootful: true,
-        startNow: true,
-        setAsDefault: false,
-        virtualizationProvider: getVirtualizationProvider(),
-      });
-    });
-
-    test('Environment filter becomes visible with two running machines', async ({ navigationBar }) => {
-      const pods = await navigationBar.openPods();
-      await playExpect.poll(async () => await pods.isEnvironmentFilterVisible(), { timeout: 30_000 }).toBeTruthy();
-    });
-
-    test('Filtering pods by environment', async ({ navigationBar }) => {
-      test.setTimeout(120_000);
-
-      let images = await navigationBar.openImages();
-      const pullImagePage = await images.openPullImage();
-      images = await pullImagePage.pullImage(environmentFilterImage, 'latest', 60_000, 'podman-machine-default');
-      await playExpect
-        .poll(async () => await images.waitForImageExists(environmentFilterImage), { timeout: 10_000 })
-        .toBeTruthy();
-
-      const imageDetails = await images.openImageDetails(environmentFilterImage);
-      const runImage = await imageDetails.openRunImage();
-      const containers = await runImage.startContainer(environmentFilterContainer, containerStartParams);
-      await playExpect(containers.header).toBeVisible();
-      await playExpect
-        .poll(async () => await containers.containerExists(environmentFilterContainer), { timeout: 15_000 })
-        .toBeTruthy();
-
-      const createPodPage = await containers.openCreatePodPage([environmentFilterContainer]);
-      const pods = await createPodPage.createPod(environmentFilterPod);
-      await playExpect(pods.heading).toBeVisible({ timeout: 60_000 });
-      await playExpect.poll(async () => await pods.podExists(environmentFilterPod), { timeout: 15_000 }).toBeTruthy();
-
-      await test.step(`Filter by ${secondMachineDisplayName} shows no pods`, async () => {
-        await pods.filterByEnvironment(secondMachineDisplayName);
-        await playExpect.poll(async () => await pods.countRowsFromTable(), { timeout: 10_000 }).toBe(0);
-      });
-
-      await test.step(`Filter by ${defaultMachineDisplayName} shows the created pod`, async () => {
-        await pods.filterByEnvironment(defaultMachineDisplayName);
-        await playExpect
-          .poll(async () => await pods.countRowsFromTable(), { timeout: 10_000 })
-          .toBeGreaterThanOrEqual(1);
-        await playExpect.poll(async () => await pods.podExists(environmentFilterPod), { timeout: 10_000 }).toBeTruthy();
-      });
-
-      await pods.clearFilterByEnvironment();
     });
   });
